refactor cheat channel to clean up peers collection
